MGS-4784 Ensure that a unique url is used when creating a publication with the same title as an existing one
When importing a publications we first try to find out if it already exists by querying by isbn. If the isbn exists then
that publicaitons will be updated.  If it does not then we try to create a new one at a path determined by the
PublicationPathStategy class.  This class will return a path that appends a number to the end of the title if
the title already exists to aboid a url clash.