MGS-4724 Allocate govscot:slug property
In order to provide a unique url for a publication we are introducing the
govscot:slug property that will be used by the link processor.  This commit
ensures that we set this property with a unique value when creating a publication.

The strategy used is to append a number to the end of a sluggified version of
the title.