MGS-4784 Add support for rewriting links with anchors
Hippo does not handle links to local anchors properly. To get around this the
PublicationLinkRewriter now rewrites these links to be of the form
/publications/publicaiton-slug/pages/1#my-anchor

This means that the link will break if the publicaiton or page is moved but
this is not avoidable right now.
See: https://issues.onehippo.com/browse/HSTTWO-4538