MGS-5020 parse executive summary as markdown
When publicaitons are imported with no html the executive summary is used as the content.

Frequently this contains newlines and/or bullet points.  However the newlines are not
significant in html and so the text was not rendered as the user expects. To resolve this
we now treat the executive sumary as markdown.