MGS-5633 fix for duplicate slugs bug
The query ised to find existing publications for a given slug
was being used wrongly - the code expected there to be exactly
one result when it is possible for there to be more in the case
the publication has been edited.  Changed to check for at least
one result instead.