MGS-7910 Fix test discrepancy with timestamp resolution
We observed differences in timestamp resolution with the Java 17 upgrade when
running these tests. On macOS, timestamps appear to be truncated to
microseconds. We observed higher timestamp resolution on linux. To resolve
this, we decided to truncate the timestamp in the test.

The default precision in H2 is to 6 fractional seconds, and can be increased up
to 9. However, PostgreSQL reduces any greater precision back to 6, meaning a
migration wouldn't help.

Also removed an unused test.